requirements.txt
duckdb>=0.9.0
pandas>=2.0.0
pyarrow>=14.0.0
requests>=2.31.0
tenacity>=8.2.0
python-dotenv>=1.0.0
pydantic>=2.5.0
geopandas>=0.14.0
shapely>=2.0.0
pyproj>=3.6.0
streamlit>=1.29.0
plotly>=5.18.0
jsonschema>=4.20.0
camelot-py[cv]>=0.11.0
tabula-py>=2.9.0
matplotlib>=3.8.0
fiona>=1.9.0
rtree>=1.1.0
pytest>=7.4.0
